chore: update folder structure, update readmes and tidy old files.
