+++ testing/test_deployment/README.md
@@ -1,19 +1,5 @@
-# src
+# Test Deployment
 
-This [main.tf](./main.tf) illustrates how to call this Terraform module from the source code, rather than replying on a download of the module from the Terraform Website:
+This Terraform configuration deploys the Playground to Snowflake in a way that can then be used for testing.
 
-```HCL
-module "playground" {
-  source = "../.." #"jagoodhand/playground/snowflake"
-
-  tasks_enabled = true
-  dry_run       = false
-
-  playground = {
-    database              = "PLAY"
-    schema                = "GROUND"
-    is_transient          = false
-    administration_schema = "ADMINISTRATION"
-  }
-}
-```
+This is used by the Integration tests.
