+++ testing/test_objects/README.md
@@ -1,0 +1,3 @@
+# Test Objects
+
+This Terraform Configuration creates objects that can be dropped using the Tidy Playground procedure as part of the integration testing to confirm the Playground performs as expected.
