+++ testing/tests/data/object_ages.csv
@@ -1,0 +1,13 @@
+OBJECT_DATABASE,OBJECT_SCHEMA,OBJECT_NAME,OBJECT_TYPE,OBJECT_DOMAIN,TAG_DOMAIN,SQL_OBJECT_TYPE,DAYS_SINCE_CREATION,DAYS_SINCE_LAST_ALTERATION,EXPIRY_DATE,OBJECT_OWNER
+TEST_PLAYGROUND,GROUND,"""STAGE_1""",INTERNAL_NAMED,STAGE,,STAGE,60,0,,SYSADMIN
+TEST_PLAYGROUND,GROUND,"""TASK_1""",TASK,TASK,,TASK,31,,,SYSADMIN
+TEST_PLAYGROUND,GROUND,"""VIEW_2""",VIEW,TABLE,,VIEW,10,0,9999-12-31,SYSADMIN
+TEST_PLAYGROUND,GROUND,"""VIEW_1""",VIEW,TABLE,,VIEW,7,0,,SYSADMIN
+TEST_PLAYGROUND,GROUND,"""TABLE_3""",BASE_TABLE,TABLE,,TABLE,0,0,,ACCOUNTADMIN
+TEST_PLAYGROUND,GROUND,"""STREAM_1""",STREAM,STREAM,,STREAM,50,,9999-12-31,SYSADMIN
+TEST_PLAYGROUND,GROUND,"""VIEW_5""",VIEW,TABLE,,VIEW,80,0,9999-12-31,SYSADMIN
+TEST_PLAYGROUND,GROUND,"""TABLE_2""",BASE_TABLE,TABLE,,TABLE,5,0,,SYSADMIN
+TEST_PLAYGROUND,GROUND,"""PROC_1""(VARCHAR,DATE)",PROCEDURE,PROCEDURE,,PROCEDURE,32,0,,SYSADMIN
+TEST_PLAYGROUND,GROUND,"""VIEW_4""",VIEW,TABLE,,VIEW,70,0,,SYSADMIN
+TEST_PLAYGROUND,GROUND,"""TABLE_1""",BASE_TABLE,TABLE,,TABLE,3,0,,SYSADMIN
+TEST_PLAYGROUND,GROUND,"""VIEW_6""",VIEW,TABLE,,VIEW,90,0,,SYSADMIN
