+++ testing/tests/pyproject.toml
@@ -1,0 +1,18 @@
+[tool.poetry]
+name = "test-script"
+version = "0.1.0"
+description = ""
+authors = ["James Goodhand <james.goodhand@hey.com>"]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = "^3.8"
+pytest = "^7.2.2"
+snowflake-connector-python = {extras = ["pandas"], version = "^3.0.1"}
+cryptography = "^39.0.2"
+snapshottest = "^0.6.0"
+
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
